+++ src_tests/arm_neon/test_vceq.cpp
@@ -7,30 +7,30 @@ using namespace iris;
 template <typename T, typename R>
 void test_vceq(R(*func)(T,T)) {
     T v1, v2;
-    for(int i = 0; i < T::length; i++) {
+    for(size_t i = 0; i < T::length; i++) {
         v1.template at<typename T::elementType>(i) = i;
         v2.template at<typename T::elementType>(i) = i;
     }
     R result = func(v1,v2);
-    for(int i = 0; i < R::length; i++) {
+    for(size_t i = 0; i < R::length; i++) {
         assert(result.template at<typename R::elementType>(i) == std::numeric_limits<typename R::elementType>::max());
     }
 
-    for(int i = 0; i < T::length; i++) {
+    for(size_t i = 0; i < T::length; i++) {
         v1.template at<typename T::elementType>(i) = i;
         v2.template at<typename T::elementType>(i) = i+1;
     }
     result = func(v1,v2);
-    for(int i = 0; i < T::length; i++) {
+    for(size_t i = 0; i < T::length; i++) {
         assert(result.template at<typename R::elementType>(i) == std::numeric_limits<typename R::elementType>::min());
     }
 
-    for(int i = 0; i < T::length; i++) {
+    for(size_t i = 0; i < T::length; i++) {
         v1.template at<typename T::elementType>(i) = i+1;
         v2.template at<typename T::elementType>(i) = i;
     }
     result = func(v1,v2);
-    for(int i = 0; i < T::length; i++) {
+    for(size_t i = 0; i < T::length; i++) {
         assert(result.template at<typename R::elementType>(i) == std::numeric_limits<typename R::elementType>::min());
     }
 }
