+++ src_tests/arm_neon/test_vld1.cpp
@@ -6,11 +6,11 @@ using namespace iris;
 template<typename T>
 void test_vld1(T(*func)(typename T::elementType*)) {
 	typename T::elementType data[T::length];
-	for (int i = 0; i < T::length; i++) {
+	for (size_t i = 0; i < T::length; i++) {
 		data[i] = i;
 	}
 	T output = func(data);
-	for (int i = 0; i < T::length; i++) {
+	for (size_t i = 0; i < T::length; i++) {
 		bool cmp = output.template at<typename T::elementType>(i) == data[i];
 		assert(cmp);
 	}
