+++ src_tests/arm_neon/test_vmax.cpp
@@ -6,12 +6,12 @@ using namespace iris;
 template <typename T>
 void test_vmax(T(*func)(T,T)) {
     T v1, v2;
-    for(int i = 0; i < T::length; i++) {
+    for(size_t i = 0; i < T::length; i++) {
         v1.template at<typename T::elementType>(i) = i;
         v1.template at<typename T::elementType>(i) = T::length - i;
     }
     T result = func(v1,v2);
-    for(int i = 0; i < T::length/2; i++) {
+    for(size_t i = 0; i < T::length/2; i++) {
         assert(result.template at<typename T::elementType>(i) >= v1.template at<typename T::elementType>(i));
         assert(result.template at<typename T::elementType>(i) >= v1.template at<typename T::elementType>(i));
     }
