+++ src_tests/arm_neon/test_vqadd.cpp
@@ -9,12 +9,12 @@ using namespace iris;
 template<typename T>
 void test_vqadd(T(*func)(T,T)) {
     T v1, v2;
-    for(int i = 0; i < T::length; i++) {
+    for(size_t i = 0; i < T::length; i++) {
         v1.template at<typename T::elementType>(i) = std::numeric_limits<typename T::elementType>::max() - i;
         v2.template at<typename T::elementType>(i) = i + 1;
     }
     T result = func(v1,v2);
-    for(int i = 0; i < T::length; i++) {
+    for(size_t i = 0; i < T::length; i++) {
         if(std::is_signed<typename T::elementType>::value) {
             std::cout << "Expected: " << static_cast<int64_t>(std::numeric_limits<typename T::elementType>::max())
                       << ", actual: " << static_cast<int64_t>(result.template at<typename T::elementType>(i)) << std::endl;
