+++ src_tests/arm_neon/test_vsubl.cpp
@@ -9,15 +9,15 @@ using namespace iris;
 template<typename T, typename R>
 void test_vsubl(R(*func)(T,T)) {
     T v1, v2;
-    for(int i = 0; i < T::length; i++) {
+    for(size_t i = 0; i < T::length; i++) {
         v1.template at<typename T::elementType>(i) = i;
         v2.template at<typename T::elementType>(i) = i + 1;
     }
     R result = func(v2,v1);
-    for(int i = 0; i < R::length; i++) {
+    for(size_t i = 0; i < R::length; i++) {
         assert(result.template at<typename R::elementType>(i) == 1);
     }
-    for(int i = 0; i < T::length; i++) {
+    for(size_t i = 0; i < T::length; i++) {
         v1.template at<typename T::elementType>(i) = std::numeric_limits<typename T::elementType>::max();
         if(std::is_signed<typename T::elementType>::value) {
             v2.template at<typename T::elementType>(i) = -i;
@@ -26,7 +26,7 @@ void test_vsubl(R(*func)(T,T)) {
         }
     }
     result = func(v1,v2);
-    for(int i = 0; i < R::length; i++) {
+    for(size_t i = 0; i < R::length; i++) {
         typename R::elementType value;
         if(std::is_signed<typename T::elementType>::value) {
             value = static_cast<typename R::elementType>(std::numeric_limits<typename T::elementType>::max()) + i;
