+++ src_tests/arm_neon/test_vsubw.cpp
@@ -6,15 +6,15 @@ template<typename R, typename T>
 void test_vsubw(R(*func)(R,T)) {
     R v1;
     T v2;
-    for(int i = 0; i < T::length; i++) {
+    for(size_t i = 0; i < T::length; i++) {
         v1.template at<typename R::elementType>(i) = i + 1;
         v2.template at<typename T::elementType>(i) = i;
     }
     R result = func(v1,v2);
-    for(int i = 0; i < R::length; i++) {
+    for(size_t i = 0; i < R::length; i++) {
         assert(result.template at<typename R::elementType>(i) == 1);
     }
-    for(int i = 0; i < T::length; i++) {
+    for(size_t i = 0; i < T::length; i++) {
         v1.template at<typename R::elementType>(i) = static_cast<typename R::elementType>(std::numeric_limits<typename T::elementType>::max());
         if(std::is_signed<typename T::elementType>::value) {
             v2.template at<typename T::elementType>(i) = -i;
@@ -23,7 +23,7 @@ void test_vsubw(R(*func)(R,T)) {
         }
     }
     result = func(v1,v2);
-    for(int i = 0; i < R::length; i++) {
+    for(size_t i = 0; i < R::length; i++) {
         typename R::elementType value;
         if(std::is_signed<typename T::elementType>::value) {
             value = static_cast<typename R::elementType>(std::numeric_limits<typename T::elementType>::max()) + i;
